services/emotion-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
python-dotenv==1.0.0
anthropic==0.18.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
